+++ crossover/__main__.py
@@ -1,9 +1,11 @@
 import logging
 
+from crossover.crossers import Crossers
 from database_handler.handlers import DatabaseHandlers
 from database_handler.redis_handler import RedisHandler
+from message_handler.handlers import MessageHandlers
 from message_handler.rabbit_message_queue import RabbitMessageQueue
-from utilities.utils import MessageHandlers, Crossers, __set_crosser, get_pga_id, set_property
+from utilities.utils import __set_crosser, get_pga_id, set_property
 
 logging.basicConfig(level=logging.DEBUG)  # TODO: remove and reduce to INFO
